backend/requirements.txt
Django==5.2.1
django-cors-headers==4.7.0
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular[swagger-ui]==0.28.0
inflection==0.5.1
jsonschema==4.23.0
psycopg2-binary==2.9.10
PyJWT==2.9.0
python-dotenv==1.1.0
PyYAML==6.0.2
uritemplate==4.1.1
python-dateutil